Truncate decimals when querying binance

+++ app/analysis.py
@@ -32,7 +32,7 @@ class StrategyAnalyzer():
     def get_historical_data(self, market_pair, exchange, time_unit, max_days=100):
         # The data_start_date timestamp must be in milliseconds hence * 1000.
         data_start_date = datetime.now() - timedelta(days=max_days)
-        data_start_date = data_start_date.replace(tzinfo=timezone.utc).timestamp() * 1000
+        data_start_date = int(data_start_date.replace(tzinfo=timezone.utc).timestamp() * 1000)
         historical_data = self.__exchange_interface.get_historical_data(
             market_pair=market_pair,
             exchange=exchange,
